# apps/front/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/cargo-documents": {
      "get": {
        "operationId": "list_cargo_documents",
        "tags": [
          "cargo_documents"
        ],
        "description": "업로드된 화물 문서 메타데이터 목록",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargoDocumentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "upload_cargo_document",
        "tags": [
          "cargo_documents"
        ],
        "description": "B/L·DGD 원문 업로드. 파일은 로컬 스토리지에 UUID 이름으로 저장하고 메타데이터만 DB에 기록한다.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadCargoDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/cargo-documents/{id}": {
      "get": {
        "operationId": "get_cargo_document",
        "tags": [
          "cargo_documents"
        ],
        "description": "업로드된 화물 문서 메타데이터 단건 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/cargo-documents/{id}/confirm-review": {
      "post": {
        "operationId": "confirm_cargo_document_review",
        "tags": [
          "cargo_documents"
        ],
        "description": "화물 문서 검수 확정 (FR-B2, 데모 시나리오 6)\nMSDS 원문을 관리자가 확인한 뒤 CONFIRMED로 확정한다.\n이미 CONFIRMED인 문서도 멱등하게 200을 반환한다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CargoDocumentResponse": {
        "type": "object",
        "properties": {
          "cargo_document_id": {
            "type": "integer",
            "format": "int64"
          },
          "content_type": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "document_type": {
            "type": "string"
          },
          "file_format": {
            "type": "string"
          },
          "file_hash": {
            "type": "string",
            "nullable": true
          },
          "file_size": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "file_url": {
            "type": "string"
          },
          "original_file_name": {
            "type": "string",
            "nullable": true
          },
          "review_status": {
            "type": "string",
            "description": "PENDING | CONFIRMED"
          },
          "uploaded_by_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "cargo_document_id",
          "document_type",
          "file_format",
          "file_url",
          "review_status",
          "uploaded_by_id",
          "created_at"
        ]
      },
...
      "UploadCargoDocumentRequest": {
        "type": "object",
        "properties": {
          "document_type": {
            "type": "string",
            "description": "BL | DGD"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "document_type",
          "file"
        ]
      },
...
    }
  },
...
}